requirements.txt
# SparX — voice node, NemoClaw, FormFinder, scrapers
#
# Install order (important):
#
#   1) Matching torch + torchaudio (same CUDA build), e.g.:
#        pip install -r requirements-torch-cu130.txt
#      or requirements-torch-cu124.txt — match pytorch.org to your driver.
#
#   2) Then everything else:
#        pip install -r requirements.txt
#
# If `import torchaudio` fails with:
#   undefined symbol: ... fft_irfft ...
# your torch and torchaudio came from different builds. Fix:
#   pip uninstall -y torch torchaudio
#   pip install -r requirements-torch-cu124.txt
#
# System: ffmpeg (for pydub), openssl (voice_node self-signed certs)
#
# If `import torchaudio` fails with:
#   OSError: libcudart.so.12: cannot open shared object file
#   pip install nvidia-cuda-runtime-cu12
#   source sparx/scripts/source_cuda_libs.sh

# --- Voice server & HTTP ---
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pydub>=0.25.0
httpx>=0.24.0
pydantic>=2.0.0
requests>=2.28.0

# --- Parakeet / NeMo ASR (asr_engine.py) ---
# libcudart for GPU wheels when no full system CUDA 12 on LD_LIBRARY_PATH
nvidia-cuda-runtime-cu12>=12.1.0
# torch + torchaudio: install from requirements-torch-*.txt first (not plain PyPI).
nemo_toolkit[asr]>=1.21.0

# --- CLI mic / file ASR helpers in asr_engine.py ---
sounddevice>=0.4.6
soundfile>=0.12.0

# --- form_finder/nyc311_ka.py ---
beautifulsoup4>=4.12.0
